# openapi.yaml
openapi: 3.0.0
info:
  title: Notizen API
  description: Eine API zur Verwaltung von Notizen mit Flask
  version: 1.0.0
servers:
  - url: http://localhost:5000/
paths:
  /api/call/:
    get:
      summary: Get all notes of a user
      parameters:
        - in: query
          name: username
          description: The email of the user
          required: true
          schema:
            type: string
        - in: query
          name: password
          description: The password of the user
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  Page:
                    type: string
                    example: API
                  Status:
                    type: integer
                    example: 200
                  Message:
                    type: string
                    example: OK
                  Notizen:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        data:
                          type: string
                          example: "Lorem ipsum"
                        date:
                          type: string
                          format: date-time
                          example: "2023-05-01 10:00:00"
        '204':
          description: No Content
          content:
            application/json:
              schema:
                type: object
                properties:
                  Page:
                    type: string
                    example: API
                  Status:
                    type: integer
                    example: 204
                  Message:
                    type: string
                    example: No Content
                  error:
                    type: string
                    example: Keine Notizen vorhanden
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  Page:
                    type: string
                    example: API
                  Status:
                    type: integer
                    example: 401
                  Message:
                    type: string
                    example: Unauthorized
                  error:
                    type: string
                    example: Username-Passwort-Kombination existiert nicht
  /api/create/:
    post:
      summary: Create a new note
      description: Create a new note for a user with the provided username and password
      parameters:
        - in: query
          name: username
          required: true
          schema:
            type: string
          description: The username of the user who will own the note
        - in: query
          name: password
          required: true
          schema:
            type: string
          description: The password of the user who will own the note
        - in: query
          name: note
          required: true
          schema:
            type: string
          description: The content of the note to create
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  Page:
                    type: string
                    example: API
                  Status:
                    type: integer
                    example: 201
                  Message:
                    type: string
                    example: Created
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  Page:
                    type: string
                    example: API
                  Status:
                    type: integer
                    example: 400
                  Message:
                    type: string
                    example: Bad Request
                  error:
                    type: string
                    example: Notiz ist leer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  Page:
                    type: string
                    example: API
                  Status:
                    type: integer
                    example: 401
                  Message:
                    type: string
                    example: Unauthorized
                  error:
                    type: string
                    example: Username-Passwort-Kombination existiert nicht
  /api/delete/:
    get:
      summary: Delete all notes
      description: Delete all notes for a user with the provided username and password
      parameters:
        - in: query
          name: username
          required: true
          schema:
            type: string
          description: The username of the user whose notes to delete
        - in: query
          name: password
          required: true
          schema:
            type: string
          description: The password of the user whose notes to delete
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  Page:
                    type: string
                    example: API
                  Status:
                    type: integer
                    example: 200
                  Message:
                    type: string
                    example: OK
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  Page:
                    type: string
                    example: API
                  Status:
                    type: integer
                    example: 401
                  Message:
                    type: string
                    example: Unauthorized   
  /api/delete-recent/:
    get:
      summary: Löscht die letzte Notiz eines Benutzers.
      parameters:
        - name: username
          in: query
          required: true
          description: Der Benutzername des Benutzers.
          schema:
            type: string
        - name: password
          in: query
          required: true
          description: Das Passwort des Benutzers.
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  Page:
                    type: string
                    example: API
                  Status:
                    type: integer
                    example: 200
                  Message:
                    type: string
                    example: OK
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  Page:
                    type: string
                    example: API
                  Status:
                    type: integer
                    example: 400
                  Message:
                    type: string
                    example: Bad Request
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  Page:
                    type: string
                    example: API
                  Status:
                    type: integer
                    example: 401
                  Message:
                    type: string
                    example: Unauthorized
                  error:
                    type: string
                    example: Username-Passwort-Kombination existiert nicht
        "406":
          description: Not Acceptable
          content:
            application/json:
              schema:
                type: object
                properties:
                  Page:
                    type: string
                    example: API
                  Status:
                    type: integer
                    example: 406
                  Message:
                    type: string
                    example: Not Acceptable
                  error:
                    type: string
                    example: Keine Notizen vorhanden